seconde

+++ src/main/java/tn/esprit/spring/repositories/ReclamationRepository.java
@@ -2,6 +2,12 @@ package tn.esprit.spring.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.entities.Reclamation;
+import tn.esprit.spring.services.Stat;
 
-public interface ReclamationRepository extends JpaRepository<Reclamation,Integer> {
+import java.util.List;
+
+public interface ReclamationRepository extends JpaRepository<Reclamation,Long> {
+
+
+
 }
